new changes

+++ programs/raxml.py
@@ -254,30 +254,26 @@ def remove_redundant_sequences(curr_run_directory, prefix,msa_path, model, out_m
         if prefix in file:
             os.remove(os.path.join(curr_run_directory,file))
 
-def raxml_bootstrap_pipeline(curr_run_directory, results_folder, msa_path, prefix, model, n_cpus = 1, n_workers ='auto', redo = True):
+def raxml_bootstrap_pipeline(curr_run_directory, results_folder, msa_path, prefix, model, n_cpus = 1, n_workers ='auto', use_existing_trees = False):
 
     search_prefix = os.path.join(curr_run_directory, prefix)
     search_command = (
-        "{raxml_exe} --blopt nr_safe --force perf_threads --all --bs-trees autoMRE --threads {n_cpus} --workers {n_workers}  --force msa --force perf_threads --msa {msa_path} --model {model}  --seed {seed} --prefix {prefix}").format(raxml_exe=RAXML_NG_EXE,
+        "{raxml_exe} --blopt nr_safe --force perf_threads --all --bs-trees autoMRE --threads {n_cpus} --workers {n_workers}  --force msa --force perf_threads --msa {msa_path} --model {model}  --seed {seed} --prefix {prefix} --redo").format(raxml_exe=RAXML_NG_EXE,
         msa_path=msa_path,  seed=SEED,
         prefix=search_prefix, model=model, n_cpus = n_cpus, n_workers = n_workers)
-    if redo:
-        search_command+=' --redo'
     raxml_log_file = search_prefix + ".raxml.log"
-    model_file = search_prefix+".raxml.bestModel"
-    execute_command_and_write_to_log(search_command, print_to_log=True)
-    with open(model_file,'r') as MODEL:
-        model_str = MODEL.read().split(',')[0]
-    bootstrap_run_time = extract_bootstrap_running_time(raxml_log_file)
+    final_tree_topology_path = os.path.join(results_folder,'raxml_final_tree_topology.tree')
+    all_final_tree_topologies_path = os.path.join(results_folder, 'raxml_all_final_tree_topologies.tree')
     best_tree_topology_path_orig = search_prefix + ".raxml.support"
     all_final_trees_path_orig = search_prefix + ".raxml.mlTrees"
+    if not use_existing_trees:
+        logging.info("Re-estimating final trees")
+        execute_command_and_write_to_log(search_command, print_to_log=True)
+        shutil.move(best_tree_topology_path_orig, final_tree_topology_path)
 
-    #model_str = re.sub('\+FU\{[^{}]*\}', '', model_str)
-    #model_str = model_str.replace('4m', '')
-    final_tree_topology_path = os.path.join(results_folder,'raxml_final_tree_topology.tree')
-    shutil.move(best_tree_topology_path_orig,final_tree_topology_path)
-    all_final_tree_topologies_path = os.path.join(results_folder, 'raxml_all_final_tree_topologies.tree')
-    shutil.move(all_final_trees_path_orig,all_final_tree_topologies_path)
+        shutil.move(all_final_trees_path_orig, all_final_tree_topologies_path)
+
+    bootstrap_run_time = extract_bootstrap_running_time(raxml_log_file)
     res = {
            'final_tree_topology_path': final_tree_topology_path, 'all_final_tree_topologies_path':all_final_tree_topologies_path,'program_boot_run_time':bootstrap_run_time}
     return res
